+++ Utils/KatoMethodRequirement.h
@@ -26,7 +26,7 @@ namespace Requirement {
 		
 
 		//加藤さん修論手法のみに必要なパラメータ
-		int max_variation_of_pause_time;
+		double max_variation_of_pause_time;
 		double max_variation_of_speed;
 
 		
@@ -50,7 +50,7 @@ namespace Requirement {
 			double required_anonymous_area,
 			size_t dummy_num,
 			int service_interval,
-			int max_variation_of_pause_time,
+			double max_variation_of_pause_time,
 			double max_variation_of_speed,
 			int interval_of_base_phase,
 			int cycle_of_interval_of_base_phase,
